+++ cli/tests/testdata/npm/local_dir_resolves_symlinks/index.js
@@ -1,0 +1,3 @@
+import * as d from "define-properties";
+
+console.log(typeof d.default === "function", "it works");
