+++ cli/tests/testdata/npm/local_dir_resolves_symlinks/index.out
@@ -1,0 +1,2 @@
+Download [WILDCARD]
+true it works
