+++ cli/tests/testdata/npm/local_dir_resolves_symlinks/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "foo",
+  "type": "module",
+  "dependencies": {
+    "define-properties": "^1.2.0"
+  }
+}
